:bomb: refactor routes into unique router.js
