+++ src/api/provinces.js
@@ -1,9 +1,6 @@
-const express = require('express');
 const { Province } = require('../../db/models');
 
-const router = express.Router();
-
-const getAll = async (req, res, next) => {
+const allProvinces = async (req, res, next) => {
   try {
     const info = await Province.findAll({
       attributes: { exclude: ['createdAt', 'updatedAt'] },
@@ -14,7 +11,4 @@ const getAll = async (req, res, next) => {
   }
 };
 
-// Routes
-router.get('/', getAll);
-
-module.exports = router;
+module.exports = { allProvinces };
