+++ src/server.js
@@ -22,10 +22,7 @@ app.use((req, res, next) => {
 });
 
 // Rutas
-app.use('/login', require('./api/login'));
-app.use('/provinces', require('./api/provinces'));
-app.use('/towns', require('./api/towns'));
-app.use('/users', require('./api/users'));
+app.use('/', require('./router'));
 
 // Not Found Handling
 app.use((req, res, next) => {
